+++ tox.ini
@@ -12,7 +12,7 @@ basepython = python3
 
 [testenv]
 commands =
-    pytest --ignore="tests/test_integration.py" {posargs:-vvs}
+    pytest --ignore=tests/integration {posargs:-vvs}
 
     # Ensure pex's main entrypoint can be run externally.
     pex --disable-cache . -e pex.bin.pex:main --version
@@ -41,7 +41,7 @@ deps =
     pytest-xdist==1.34.0
     {[testenv]deps}
 commands =
-    pytest -n auto tests/test_integration.py {posargs:-vvs}
+    pytest -n auto tests/integration {posargs:-vvs}
 
 [testenv:format-run]
 skip_install = true
